CropCanvas emits crop results after crop operation.
This streamlines cropping behaviour among widgets.

It can emit a QImage, a QPixmap or both. The last QImage from crop
operation will be held and can be accessed later. QPixmap will be held
if it is emitted or else generated when asked specifically. (we skip
QPixmap generation unless asked for, as QPixmap will only be used for
showing cropped image.)