Cropped image is scaled and centered if it does not fit right side.
If the cropped image is bigger than the right side of the splitter it is
scaled down to a size which fits that area (maintaining the aspect
ratio). It is triggered by imageCropped signal.

Now I need to do the same thing when the cropped image does not fit
after the splitter is moved or the window is resized.