Cropped image is resizable and preserves aspect ratio when it scales.
I created a new class called ARPLabel (Aspect Ratio Preserving QLabel)
and moved the functionality away from MainWindow. It has a QLabel
centered inside an HBox and a VBox. Label's sizeHint is ignored (or we
can't make the image smaller!) and maximumSize is set after resize if
necessary. This way image can be scaled down when splitter squeezes it
or MainWindow is resized down.

I first tried to do this by inserting an event filter in MainWindow. It
does not work since label does not remember the original image size and
you can't override sizeHint. (I tried to do that with setProperty but
didn't work.)